Add shared test suites to eliminate Maven/Gradle duplication
- Created shared/restSnippetSuite.ts for REST snippet tests
- Created shared/hoverTestSuite.ts for hover tests
- Created shared/devModeTestSuite.ts for DevMode tests (6 shared tests)
- Refactored Maven/Gradle entry files to use shared suites
- Achieved ~50% code reduction through parameterization
- Kept build-tool-specific tests in separate describe blocks
